Added CPU model and inference

+++ src/mouse_drawing.h
@@ -23,7 +23,7 @@ public:
   void save_screen(void);
   // Exports screen to a grayscale vector
   // takes the width and height to be scaled to
-  std::vector<double> export_to_vector(int w, int h, double scale);
+  std::vector<float> export_to_vector(int w, int h, double scale);
 
   virtual ~MouseDrawing();
   // State definitions to decide how to handle on_draw calls
